Pass puz handle through application and refresh list "properly"
Pass through application since it's how it's done for notes and
cluelistactivity too.

Refresh is now complicated. Previously it saved the last tapped view,
which isn't really safe afaik as the recycler view can reuse views.

Instead the puzzle list adapter is searched for one with the same file,
it is updated and the adapter notified to update that element.